Managing product page

+++ QIMA_BE/src/main/java/com/qima/controllers/ProductController.java
@@ -37,7 +37,7 @@ public class ProductController {
     }
 
     @PutMapping("/{id}")
-    public ProductDTO updateProduct(@PathVariable Integer id, @RequestBody ProductDTO productDTO) {
+    public ProductDTO updateProduct(@PathVariable Long id, @RequestBody ProductDTO productDTO) {
         productDTO.setId(id);
         return productMapper.toDTO(productRepository.save(productMapper.toEntity(productDTO)));
     }
